backend/requirements.txt
# Sentinel - Complete Requirements for Render Deployment
# All necessary dependencies for full functionality

# === Core Backend ===
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Limiter>=3.5.0
gunicorn>=21.2.0
requests>=2.31.0
python-dotenv>=1.0.0

# === Essential Data Processing ===
pandas>=2.0.0
numpy>=1.24.0

# === Machine Learning ===
scikit-learn>=1.3.0
torch>=2.1.0

# === Database ===
psycopg2-binary>=2.9.0


# === Scheduling & Background Tasks ===
APScheduler>=3.10.0

# === Multi-Chain Support ===
web3>=6.12.0
base58>=2.1.0

# === API & Security ===
marshmallow>=3.20.0
cryptography>=41.0.0
pydantic>=2.0.0
Werkzeug>=2.3.0

# === HTTP & Async Clients ===
httpx>=0.24.0
aiohttp>=3.8.0

# === GraphQL Client ===
gql>=3.4.0

# === Data Validation & Serialization ===
jsonschema>=4.17.0

# === Utilities ===
python-dateutil>=2.8.0
pytz>=2023.3

# === Production Monitoring ===
prometheus-client>=0.19.0

# === Database ORM ===
SQLAlchemy>=2.0.0

# === Testing & Development (Optional) ===
pytest>=7.4.0 